requirements.txt
Django==1.7.4
django-allauth==0.19.1
django-crispy-forms==1.4.0
django-extensions==1.5.0
oauthlib==0.7.2
psycopg2==2.6
python-openid==2.2.5
requests==2.5.3
requests-oauthlib==0.4.2
six==1.9.0
stripe==1.21.0
wsgiref==0.1.2
